Changed nloptr.default.options to function.
Previously, nloptr.default.options was a data.frame. As this data was
not exported it results in errors when another package calls nloptr
(which uses nloptr.default.options) while the nloptr package is not
attached. The same data.frame is now return from the exported function
nloptr.get.default.options in order to solve this problem.