Added advance warning for future change
Inequalities in functions that came from nloptwrap (auglag, cobyla,
isres, mma, and slsqp) are now defined differently than in the rest of
the nloptr packages (they use >= instead of <=). To remove this
inconsistency I will change this in a next version of nloptr. In the
currenct version a warning is added to these functions as an advance
warning when using an inequality constraints.