Moved version check in configure
Moved the check tot test whether the version of NLopt is 2.4.0 or
greater to occur after the check whether the NLopt header is found. In
the previous version of the configure file, if the version was
incorrect, but a header was found, then the pre-installed NLopt would
still be used, and result in an error during installation of nloptr, as
the version of NLopt was too low.