Expose NLopt C functions through API
C functions in NLopt can now be called directly from other R packages:
   inst/include/nloptrAPI.h
   src/init_nloptr.c

A test to call NLopt C functions is provided using an example that uses the inline package:
   tests/testthat/test-C-api.R

Two types of compiler and linker flags are defined in configure. Flags that are related to compiling and building the nloptr library during installation of the package. Flags that can be used when linking to the NLopt library through nloptr from R:
   R/PkgFlags.R
   R/PkgFlags.R.in
   configure
   configure.ac
   src/Makevars.in